# bookingsvc/api/openapi/booking.openapi.yaml
openapi: 3.0.3

info:
  title: Booking Service API
  version: 1.0.0

servers:
  - url: /api/v1

paths:
  /hotels/{hotel_id}/available-rooms:
    get:
      summary: Get available rooms for hotel by date range
      operationId: getAvailableRooms
      parameters:
        - name: hotel_id
          in: path
          required: true
          schema:
            type: integer
            format: int64

        - name: date_from
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2026-05-01"

        - name: date_to
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2026-05-05"

      responses:
        "200":
          description: Available rooms list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableRoomsResponse"

        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings:
    get:
      summary: Get bookings
      operationId: getBookings
      parameters:
        - name: user_id
          in: query
          required: false
          schema:
            type: integer
            format: int64

        - name: hotel_id
          in: query
          required: false
          schema:
            type: integer
            format: int64

      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingsResponse"

        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create booking
      operationId: createBooking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingRequest"

      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"

        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "409":
          description: Room already booked for selected dates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bookings/{booking_id}:
    get:
      summary: Get booking by id
      operationId: getBookingByID
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: integer
            format: int64

      responses:
        "200":
          description: Booking info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingResponse"

        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      summary: Cancel booking
      operationId: cancelBooking
      parameters:
        - name: booking_id
          in: path
          required: true
          schema:
            type: integer
            format: int64

      responses:
        "204":
          description: Booking cancelled

        "404":
          description: Booking not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    Room:
      type: object
      required:
        - id
        - hotel_id
        - room_num
        - type
        - cost
        - is_available
      properties:
        id:
          type: integer
          format: int64
        hotel_id:
          type: integer
          format: int64
        room_num:
          type: integer
          format: int64
        type:
          type: string
          example: "standard"
        cost:
          type: integer
          format: int64
        is_available:
          type: boolean

    AvailableRoomsResponse:
      type: object
      required:
        - rooms
      properties:
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/Room"

    CreateBookingRequest:
      type: object
      required:
        - user_id
        - hotel_id
        - room_id
        - date_from
        - date_to
      properties:
        user_id:
          type: integer
          format: int64
        hotel_id:
          type: integer
          format: int64
        room_id:
          type: integer
          format: int64
        date_from:
          type: string
          format: date
          example: "2026-05-01"
        date_to:
          type: string
          format: date
          example: "2026-05-05"

    BookingsResponse:
      type: object
      required:
        - bookings
      properties:
        bookings:
          type: array
          items:
            $ref: "#/components/schemas/BookingResponse"

    BookingResponse:
      type: object
      required:
        - id
        - user_id
        - hotel_id
        - room_id
        - date_from
        - date_to
        - status
      properties:
        id:
          type: integer
          format: int64
        user_id:
          type: integer
          format: int64
        hotel_id:
          type: integer
          format: int64
        room_id:
          type: integer
          format: int64
        date_from:
          type: string
          format: date
        date_to:
          type: string
          format: date
        status:
          type: string
          example: "active"

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string